Refactor structures.
- Rename `fukuroubot.py` to `bot.py`
- Move './configs/systemconfig.py` to `./config.py`
- Refactor codes that are related from the above chanages.
