+++ fukurou/__main__.py
@@ -2,8 +2,9 @@ import sys
 import logging
 import logging.config
 
-from .configs import SystemConfig, NewConfigInterrupt
-from .fukuroubot import FukurouBot
+from .configs import NewConfigInterrupt
+from .config import BotConfig
+from .bot import FukurouBot
 
 CONFIG_CREATED_MESSAGE = """ERROR: Config not found.
 
@@ -13,12 +14,12 @@ Run the program again after modifying it."""
 if __name__ == '__main__':
     # Load system config
     try:
-        SystemConfig().load(interrupt_new=True)
+        BotConfig().load(interrupt_new=True)
     except NewConfigInterrupt:
         print(CONFIG_CREATED_MESSAGE)
         sys.exit()
 
     # Load logging config
-    logging.config.dictConfig(SystemConfig().logging)
+    logging.config.dictConfig(BotConfig().logging)
 
     FukurouBot().run()
