+++ fukurou/configs/__init__.py
@@ -1,3 +1,2 @@
 from .baseconfig import BaseConfig
-from .systemconfig import SystemConfig
 from .exceptions import NewConfigInterrupt
